[Bugfix] fix race condition that leads to wrong order of token returned
During the startup of the api server the setup function is called
multiple times (every 5s). So the longer the longer the startup time
(generally for larger models) the more consumers are contending for the
output. This can then lead to race condition where the order of the
answer token is wrong.

Introduce here: https://github.com/vllm-project/vllm/pull/9973

References:
https://github.com/vllm-project/vllm/issues/10376
https://github.com/vllm-project/vllm/issues/10589
https://github.com/vllm-project/vllm/pull/10782

Signed-off-by: Jannis Schönleber <joennlae@gmail.com>